feat(all): Refactor pysg for maximum mypy compilance
BREAKING CHANGE:
